update readme with setup instructions
As a note, until this point I've been `git commit --amend`ing on
the initial commit. I know this is bad practice, but was doing
so to the end that I firmly believe in having commits be atomic
units of work (iterable of course, but definitive). When
developing a side-(side)-project like this, I've been forced to
code this in an extremely hodgepodge fashion - five minutes here
on this, twenty there on that - and I have still not reached a
point that represents any finished units of work. That said,
I now believe poor, scattered commits are better than what would
end up being one huge commit with no git history. And so,
spagetti and un-definitive as the last commit was, I'll leave it
alone and get on commiting as I would.