add ability to click on bar text to select that hour
I'm not in love with the implemention - there is no encapsulation of
hour data (the bar, the time, and the temperature), and so there are
data hooks in the dom to connect them. Ideally, each bar would be
encapsulated in a backbone view, but I'm not sure what that would look
like using d3 yet. Probable iteration here.