add feature tests for app state
Unfortunately (in respect to wanting to demo this), the state can only
be invalidly set through the url - ie, there is no way in the ui to do
so. That said, I've done validation on the url, and these are tests for
when the router sets the url to the app state model.