store data more sensibly
Finally got around to storing the data more sensibly.
All scale-relevant data are stored in english units.
Metric was considered, but because english is the default,
it is most performant (ie, it may never need converted).
Storing both was rejected out of hand for state issues.
I really wished to make use of getters and setters for the
metric data which get and set the english data.
However, Backbone and Underscores do not support object property
getters and setters (cf https://github.com/jashkenas/underscore/pull/1123)