add deploy npm script and change directory structure
I was serving out of the project directory (/weather-app--ash)
but when deployed to gh-pages the paths need to be from the root.
These changes account for this, and have the local dev server do so
as well. Finally, the /weather-app--ash/dist requirejs destination
is changed to /weather-app--ash/weather-app--ash for ease of the local
dev server in "production" mode serving out of the optimized
subdirectory. The npm deployment script moves this optimized
subdirectory up a level and deletes unnecessary cruft.